Remove contact area information destroying cast to integer in the parser.

+++ Tiles/DriverConsole/Tiles.Content.Bridge.DfNet/DfAgentBuilder.cs
@@ -607,7 +607,7 @@ namespace Tiles.Content.Bridge.DfNet
                                 IsStrike = true,
                                 IsMartialArts = true,
                                 ContactType = attack.ContactType,
-                                ContactArea = System.Math.Max(1, (int)contactArea),
+                                ContactArea = System.Math.Max(1, contactArea),
                                 MaxPenetration = System.Math.Max(1, (int)maxPen),
                                 VelocityMultiplier = 1000,
                             };
